// src/common/constants.d.ts
export declare enum ItemType {
    Node = "node",
    Edge = "edge",
    Combo = "combo"
}
export declare enum ItemState {
    Active = "active",
    ActiveAnchorPoints = "activeAnchorPoints",
    Selected = "selected",
    HighLight = "highLight",
    Error = "error"
}
export declare enum GraphType {
    Flow = "flow",
    Mind = "mind"
}
export declare enum GraphMode {
    Default = "default",
    AddNode = "addNode",
    Readonly = "readonly"
}
export declare enum GraphState {
    NodeSelected = "nodeSelected",
    EdgeSelected = "edgeSelected",
    MultiSelected = "multiSelected",
    CanvasSelected = "canvasSelected",
    ComboSelected = "comboSelected"
}
export declare enum LabelState {
    Hide = "hide",
    Show = "show"
}
export declare enum AnchorPointState {
    Enabled = "enabled",
    Disabled = "disabled"
}
export declare enum EditorEvent {
    /** 调用命令之前触发 */
    onBeforeExecuteCommand = "onBeforeExecuteCommand",
    /** 调用命令之后触发 */
    onAfterExecuteCommand = "onAfterExecuteCommand",
    /** 改变画面状态触发 */
    onGraphStateChange = "onGraphStateChange",
    /** 改变标签状态触发 */
    onLabelStateChange = "onLabelStateChange"
}